refactor: move CategoryDefinition and ReportCategoryName to all-reports.model and fix sidebar navigation
- move CategoryDefinition interface and ReportCategoryName enum from all-reports.component.ts to all-reports.model.ts
- replace hardcoded height calculation with full-viewport-height class in all-reports component
- remove inline height style from all-reports.component.scss
- fix sidebar navigation to only toggle node on click when route is not present
