+++ apps/web-giddh/src/app/all-reports/all-reports.component.html
@@ -1,5 +1,5 @@
 <div
-    class="all-reports-page bg-light-gray min-vh-100"
+    class="all-reports-page full-viewport-height bg-light-gray min-vh-100"
     appTranslate
     [file]="'all-reports'"
     (localeData)="localeData = $event"
