+++ apps/web-giddh/src/app/all-reports/utility/all-reports.model.ts
@@ -19,3 +19,31 @@ export interface ReportCategory {
     icon: string;
     reports: ReportItem[];
 }
+
+/** UI-facing shape of a report category rendered as a chip/card */
+export interface CategoryDefinition {
+    /** URL-safe key, e.g. 'cash-flow' */
+    key: string;
+    /** Human readable label from API */
+    label: string;
+    /** CSS icon class or SVG filename */
+    icon: string;
+    /** Icon foreground color */
+    color: string;
+    /** Icon badge background color */
+    bgColor: string;
+    /** Raw report `name` values that belong to this category */
+    reports: string[];
+}
+
+/** Canonical category names as returned by the API (`filterOption.name`) */
+export enum ReportCategoryName {
+    ALL = 'All',
+    SALES = 'Sales',
+    PURCHASE = 'Purchase',
+    CUSTOMERS = 'Customers',
+    FINANCIAL = 'Financial',
+    ACCOUNTING = 'Accounting',
+    INVENTORY = 'Inventory',
+    CASH_FLOW = 'CashFlow'
+}
